pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi>=0.120.0",
    "flask>=3.1.2",
    "gputil>=1.4.0",
    "psutil>=7.1.1",
    "pydantic>=2.12.3",
    "python-multipart>=0.0.20",
    "reportlab>=4.4.4",
    "requests>=2.32.5",
    "speedtest-cli>=2.1.3",
    "uvicorn>=0.38.0",
]
